fix(ltx-2.5): mirror upstream's f32 mean denominator, and gate the input that proves it
Closes the one MEDIUM finding the scoped re-review of L3's repair returned FAIL
on (issue #435, phase L3 of .agents/specs/ltx-2-5.md). The rest of that repair is
untouched and re-verified below.

THE DEFECT. `Ltx2NormAndConcatPaddedBatch` computed `denom + eps` in double.
Upstream computes it in float32, and not by choice: `sequence_lengths` is an
int64 tensor (feature_extractor.py:30), `* d` keeps it int64 (:34), and torch
promotes an int64 tensor plus a python float to the DEFAULT dtype, so the add at
:35 happens in f32. Measured against the pinned oracle at fd4ded7f:

    torch: (int64 18) + 1e-6 -> dtype=torch.float32, 18.000001907348633
    our double:               18 + 1e-6 =            18.000001000000001
    upstream mean = f32(9.0/18.000001907348633) =    0.49999994039535522
    our      mean = f32(9.0/18.000001000000001) =    0.49999997019767761

One f32 ulp. This is AGENTS.md's dtype-polarity rule in its exact stated form:
too WIDE is still numerically finer, so tokens match, every random golden passes
and nothing goes red. `range_ + eps` at :41 then multiplies that ulp by 8/eps =
8e6 whenever a (batch, layer) slice is constant, which turns it into 23842x the
suite's own kTol.

THE GATE WAS DEFINED DOWN. scripts/gen-ltx2-text-goldens.py ran upstream on
exactly the falsifying input, had the whole output tensor in `const_out`, and
reduced it to `kLtxTeNormV1ConstantSliceFinite = 1`. Both a float32 and a float64
denominator are finite, so the assertion could not see the defect. Two further
degenerate cases were reduced the same way. That is the reason this commit is
mostly test and generator: the port defect is two lines.

RED FIRST. Emitting the arrays with the dtype still unfixed:

  test_ltx2_text_encoder.cpp:938: MESSAGE: ltx2 text norm V1 constant slice max|diff| = 0.238419
  test_ltx2_text_encoder.cpp:939: ERROR: CHECK( worst < kTol ) is NOT correct!
    values: CHECK( 0.238419 <  1e-05 )
  test_ltx2_text_encoder.cpp:1002: MESSAGE: ltx2 text norm V1 near-constant max|diff| = 0.238419
  test_ltx2_text_encoder.cpp:1003: ERROR: CHECK( worst < kTol ) is NOT correct!
    values: CHECK( 0.238419 <  1e-05 )
  [doctest] test cases:   1 |   0 passed | 1 failed | 16 skipped
  [doctest] assertions: 861 | 859 passed | 2 failed |
  [doctest] Status: FAILURE!

GREEN after mirroring the denominator: those two read 0, and so does the
one-valid-token case. 17/17 cases, 3350/3350 assertions, Status: SUCCESS!

WHAT CHANGED IN THE PORT. `denom` becomes int64, as upstream's is, and the add
and divide happen in f32. The min/max become f32 too: they ARE f32 values, so
`hi - lo` is then upstream's single f32 rounding instead of an f64 subtraction
rounded twice.

The f64 SUM accumulator STAYS, and is now annotated with the measurement rather
than left bare. Upstream's `masked.sum` is an f32 reduction but a blocked one, so
no straight loop reproduces its order. Measured on this fixture against upstream:

    f64 accumulate   2.38e-07 (left)  4.77e-07 (right)
    f32 accumulate   4.77e-07 (left)  2.38e-07 (right)

A naive f32 accumulate wins on one mask and loses on the other, which is noise,
not a mirror. Accumulating exactly and rounding once is the closest
single-rounding approximation to any order, so it is the deliberate escape.

THE THREE FALSE CLAIMS. ltx2_text_encoder.h:143-148, the test's section 3 and the
generator's docstring each asserted that `denom + eps` is "UNOBSERVABLE at the
output for every possible input". It is not. It is unobservable on an all-pad row
or an all-zero mask, where :44-45 zeroes every position first. All three now say
that, name the falsifying input and its two numbers, and record that on a
realistic Gemma workload `range_` is O(1) so the same perturbation lands around
2e-12. A public header that overstates an invariant is worse than one that says
nothing, because the next reader trusts it.

Also fixes the off-by-one at ltx2_text_encoder.h:139: upstream defines `denom` at
feature_extractor.py:34 and adds `eps` at :35.

MUTATION. `denom + kEps` -> `denom + 0.0` was Status: SUCCESS before this change,
which is what the old comment's "moved nothing at all" recorded. Against the
arrays it now fails three gates, at 0.476837 (constant slice), 0.715256 (one
valid token) and 0.476837 (near-constant). Tree restored byte-for-byte
afterwards, md5 6ccebc95657cb91613e486152bc1925f.

NOT REGRESSED. The generator reproduced every pre-existing golden BYTE-IDENTICALLY
(219 insertions, 0 deletions), so the upstream anchor is intact and no prior value
drifted. The epsilon value pins are unchanged at 9.9998279390984754e-07 and
9.9999993911447867e-07. F1's `RequireDeclaredProjection` refuses by name on all
seven out-of-tree probes and the control is still accepted; A/B'd against a
library built from the pre-change source, the probe output is byte-identical.
gemma4.h is not in the diff, and its eight gates read 2/6, 2/10, 3/503, 72/161
and 1 case / 0 assertions x4, all SUCCESS.

Of the eight max|diff| values, six are unchanged and the two the reviewer
predicted improve: norm V1 (left) 4.76837e-07 -> 2.38419e-07, V1 extractor (left)
6.51926e-08 -> 5.96046e-08. Nothing got worse. docs/FEATURES.md carries the
extractor number, so it moves 6.5e-08 -> 6.0e-08.

Full gate: 391 tests, 389 passed under -j 4; test_serve_low_tools and
test_engine_core_proc failed on contention and both pass serially.

FOLLOWING_AGENTS_PROTOCOL

Following-Agents-Protocol: true
AI-Assisted: true

+++ src/vllm/model_executor/models/ltx2_text_encoder.cpp
@@ -249,26 +249,49 @@ std::vector<float> Ltx2NormAndConcatPaddedBatch(const float* stacked, const int3
     for (int64_t t = 0; t < T; ++t)
       if (mask[static_cast<size_t>(b * T + t)] != 0) ++seq_len;
     // :34 — denom = (sequence_lengths * d), i.e. the number of VALID (t, d) pairs.
-    const double denom = static_cast<double>(seq_len) * static_cast<double>(D);
+    // INT64, exactly as upstream: `attention_mask.sum(-1)` is an int64 tensor and
+    // `* d` keeps it there. The width it is added to `eps` in is decided at :35.
+    const int64_t denom = seq_len * D;
 
     for (int64_t l = 0; l < L; ++l) {
       // :33-38 — mean over the masked entries and min/max over them, both reduced
       // over BOTH the token and the hidden axis, per (batch, layer).
+      //
+      // f64 SUM ACCUMULATOR, and it is the deliberate escape rather than an
+      // oversight: upstream's `masked.sum` is a float32 reduction, but a BLOCKED
+      // one whose order no straight loop reproduces. Accumulating exactly and
+      // rounding once is the closest single-rounding approximation to any order.
+      // Measured against upstream on this fixture: f64 accumulate reads 2.38e-07
+      // (left) / 4.77e-07 (right), a naive f32 accumulate reads 4.77e-07 / 2.38e-07
+      // — it wins on one mask and loses on the other, which is noise, not a mirror.
+      // The min/max stay f32 because they ARE f32 values and `hi - lo` is then the
+      // same single rounding upstream's f32 subtraction does; taking the
+      // difference in f64 and rounding after would double-round.
       double sum = 0.0;
-      double lo = std::numeric_limits<double>::infinity();
-      double hi = -std::numeric_limits<double>::infinity();
+      float lo = std::numeric_limits<float>::infinity();
+      float hi = -std::numeric_limits<float>::infinity();
       for (int64_t t = 0; t < T; ++t) {
         if (mask[static_cast<size_t>(b * T + t)] == 0) continue;
         for (int64_t d = 0; d < D; ++d) {
-          const double v = static_cast<double>(
-              stacked[static_cast<size_t>((((b * T) + t) * D + d) * L + l)]);
-          sum += v;
+          const float v = stacked[static_cast<size_t>((((b * T) + t) * D + d) * L + l)];
+          sum += static_cast<double>(v);
           lo = std::min(lo, v);
           hi = std::max(hi, v);
         }
       }
-      const float mean = static_cast<float>(sum / (denom + kEps));
-      const float range = static_cast<float>(hi - lo);
+      // :35 — `denom + eps` where `denom` is an int64 TENSOR and `eps` a python
+      // float. Torch promotes int64-tensor + python-float to the DEFAULT dtype, so
+      // that add happens in FLOAT32, not float64: upstream's denominator for
+      // seq_len*D == 18 is 18.000001907348633, where an f64 add gives 18.000001.
+      // Adding in f64 is numerically finer and therefore WRONG here — it is a
+      // dtype that is too wide, which no token gate and no random-input golden can
+      // see. It costs one f32 ulp in `mean`, which `range_ + eps` below multiplies
+      // by 8/eps = 8e6 as `range_` collapses: on a constant slice it moves the
+      // output from 0.476837158 to 0.238418579. Gated on that input in
+      // tests/vllm/models/test_ltx2_text_encoder.cpp.
+      const float mean = static_cast<float>(sum) /
+                         (static_cast<float>(denom) + static_cast<float>(kEps));
+      const float range = hi - lo;
 
       // :41 — 8 * (x - mean) / (range + eps), applied to the UNMASKED tensor.
       for (int64_t t = 0; t < T; ++t) {
